refactor: extract Main component
